# apps/fumadocs/openapi/weather-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Weather App API",
    "version": "1.0.0",
    "description": "REST API for current weather by location, per-user search history and favourite locations.\n\nBase URL model: in production the API is served same-origin under `/api/*` via the Next.js BFF reverse proxy (session cookies stay first-party); in local development the Fastify server is reached directly at `http://localhost:3000`.\n\nAll non-2xx responses use the shared `ErrorEnvelope` shape. Authentication endpoints under `/api/auth/*` are handled by Better-Auth and are documented separately (see the Authentication page)."
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code (VALIDATION_ERROR, UNAUTHENTICATED, NOT_FOUND, LOCATION_NOT_FOUND, ALREADY_FAVOURITE, FAVOURITES_LIMIT_REACHED, FAVOURITES_OUT_OF_SYNC, UPSTREAM_ERROR, UPSTREAM_TIMEOUT, RATE_LIMITED, INTERNAL_ERROR)"
              },
              "message": {
                "type": "string",
                "description": "Human-readable, client-safe explanation"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "description": "Standard error envelope returned by every non-2xx API response."
      }
    }
  },
  "paths": {
...
    "/api/v1/favourites": {
      "get": {
        "operationId": "listFavourites",
        "summary": "List favourite locations",
        "tags": [
          "Favourites"
        ],
        "description": "Returns the signed-in user's favourite locations in display order: `sortOrder` ascending (new favourites are created at the top; PUT /api/v1/favourites/order rewrites positions), with any legacy null-`sortOrder` rows last, oldest-first. Requires a Better-Auth session cookie — see the Authentication page.",
        "responses": {
          "200": {
            "description": "The user's favourite locations, ordered as described above.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Favourite id — pass to DELETE /api/v1/favourites/{id}"
                      },
                      "name": {
                        "type": "string",
                        "description": "Place name the query resolved to"
                      },
                      "country": {
                        "type": "string",
                        "description": "ISO 3166 country code, e.g. GB"
                      },
                      "state": {
                        "description": "State/region, when the geocoder provides one",
                        "type": "string"
                      },
                      "lat": {
                        "type": "number"
                      },
                      "lon": {
                        "type": "number"
                      },
                      "sortOrder": {
                        "anyOf": [
                          {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          {
                            "type": "null"
                          }
                        ],
                        "description": "Sort position: new favourites are created above existing ones and PUT /api/v1/favourites/order rewrites positions. Ordered favourites list first; legacy null rows follow oldest-first."
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                        "description": "When the favourite was saved (ISO 8601, UTC)"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "country",
                      "lat",
                      "lon",
                      "sortOrder",
                      "createdAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid session (`UNAUTHENTICATED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (`INTERNAL_ERROR`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addFavourite",
        "summary": "Save a favourite location",
        "tags": [
          "Favourites"
        ],
        "description": "Saves a resolved location (the `location` object from GET /api/v1/weather) as a favourite, placed at the top of the list. Coordinates are the identity: saving the same `lat`/`lon` twice responds 409, and each user can hold at most 20 favourites. Requires a Better-Auth session cookie — see the Authentication page.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Place name the query resolved to"
                  },
                  "country": {
                    "type": "string",
                    "description": "ISO 3166 country code, e.g. GB"
                  },
                  "state": {
                    "description": "State/region, when the geocoder provides one",
                    "type": "string"
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lon": {
                    "type": "number"
                  }
                },
                "required": [
                  "name",
                  "country",
                  "lat",
                  "lon"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The favourite was saved; the created row is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Favourite id — pass to DELETE /api/v1/favourites/{id}"
                    },
                    "name": {
                      "type": "string",
                      "description": "Place name the query resolved to"
                    },
                    "country": {
                      "type": "string",
                      "description": "ISO 3166 country code, e.g. GB"
                    },
                    "state": {
                      "description": "State/region, when the geocoder provides one",
                      "type": "string"
                    },
                    "lat": {
                      "type": "number"
                    },
                    "lon": {
                      "type": "number"
                    },
                    "sortOrder": {
                      "anyOf": [
                        {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Sort position: new favourites are created above existing ones and PUT /api/v1/favourites/order rewrites positions. Ordered favourites list first; legacy null rows follow oldest-first."
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                      "description": "When the favourite was saved (ISO 8601, UTC)"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "country",
                    "lat",
                    "lon",
                    "sortOrder",
                    "createdAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid body (`VALIDATION_ERROR`) or the 20-favourite cap was hit (`FAVOURITES_LIMIT_REACHED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No valid session (`UNAUTHENTICATED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "This location is already a favourite (`ALREADY_FAVOURITE`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (`INTERNAL_ERROR`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/favourites/order": {
      "put": {
        "operationId": "reorderFavourites",
        "summary": "Reorder favourite locations",
        "tags": [
          "Favourites"
        ],
        "description": "Persists a manual display order. The body must contain every one of the signed-in user's favourite ids exactly once, top first; each favourite's `sortOrder` becomes its position in the list. If the id set does not match the user's current favourites (for example the list changed in another tab) the request is rejected with 409 `FAVOURITES_OUT_OF_SYNC` — re-fetch and retry. Responds with the freshly ordered list. Requires a Better-Auth session cookie — see the Authentication page.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "minItems": 1,
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "description": "Every favourite id the user currently has, in the desired display order (top first)."
                  }
                },
                "required": [
                  "ids"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The order was saved; the full reordered list is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Favourite id — pass to DELETE /api/v1/favourites/{id}"
                      },
                      "name": {
                        "type": "string",
                        "description": "Place name the query resolved to"
                      },
                      "country": {
                        "type": "string",
                        "description": "ISO 3166 country code, e.g. GB"
                      },
                      "state": {
                        "description": "State/region, when the geocoder provides one",
                        "type": "string"
                      },
                      "lat": {
                        "type": "number"
                      },
                      "lon": {
                        "type": "number"
                      },
                      "sortOrder": {
                        "anyOf": [
                          {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          {
                            "type": "null"
                          }
                        ],
                        "description": "Sort position: new favourites are created above existing ones and PUT /api/v1/favourites/order rewrites positions. Ordered favourites list first; legacy null rows follow oldest-first."
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                        "description": "When the favourite was saved (ISO 8601, UTC)"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "country",
                      "lat",
                      "lon",
                      "sortOrder",
                      "createdAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid body (`VALIDATION_ERROR`): empty, more than 20 entries, or duplicate ids.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No valid session (`UNAUTHENTICATED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "The id set does not match the user's current favourites (`FAVOURITES_OUT_OF_SYNC`) — re-fetch the list and retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (`INTERNAL_ERROR`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/favourites/{id}": {
      "delete": {
        "operationId": "deleteFavourite",
        "summary": "Remove a favourite location",
        "tags": [
          "Favourites"
        ],
        "description": "Removes one of the signed-in user's favourites. Favourites that do not exist — or belong to another user — respond 404 (ownership is never revealed). Requires a Better-Auth session cookie — see the Authentication page.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Favourite id (from GET /api/v1/favourites)"
          }
        ],
        "responses": {
          "204": {
            "description": "The favourite was removed. No body."
          },
          "400": {
            "description": "Invalid request (`VALIDATION_ERROR`): bad `id` value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "No valid session (`UNAUTHENTICATED`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No such favourite for this user (`NOT_FOUND`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (`INTERNAL_ERROR`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "This origin — proxied to the API (works on the docs site and the web app)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development (Fastify directly)"
    }
  ],
...
}