# apps/server/src/modules/favourites/favourites.routes.ts
import type { FastifyInstance, preHandlerAsyncHookHandler } from "fastify";
import type { ZodTypeProvider } from "fastify-type-provider-zod";
import { z } from "zod";

import { AppError, ErrorCodes, errorEnvelopeSchema } from "@/lib/errors";
import {
  favouriteCreateSchema,
  favouriteDeleteParamsSchema,
  favouriteItemSchema,
  favouritesListResponseSchema,
  favouritesReorderSchema,
} from "@/modules/favourites/favourites.schemas";
import type { FavouritesService } from "@/modules/favourites/favourites.service";

export interface FavouritesRoutesOptions {
  favouritesService: FavouritesService;
  requireSession: preHandlerAsyncHookHandler;
}

export async function favouritesRoutes(
  fastify: FastifyInstance,
  options: FavouritesRoutesOptions,
): Promise<void> {
  const { favouritesService, requireSession } = options;

  // Every route in this plugin requires a session.
  fastify.addHook("preHandler", requireSession);

  /** The guard always sets userId before handlers run; this narrows the type. */
  function sessionUserId(userId: string | undefined): string {
    if (!userId) {
      throw new AppError(401, ErrorCodes.UNAUTHENTICATED, "You must be signed in to do that.");
    }
    return userId;
  }

  fastify.withTypeProvider<ZodTypeProvider>().route({
    method: "GET",
    url: "/favourites",
    schema: {
      tags: ["Favourites"],
      summary: "List favourite locations",
      description:
        "Returns the signed-in user's favourite locations in display order: `sortOrder` " +
        "ascending (new favourites are created at the top; PUT /api/v1/favourites/order " +
        "rewrites positions), with any legacy null-`sortOrder` rows last, oldest-first. " +
        "Requires a Better-Auth session cookie — see the Authentication page.",
      operationId: "listFavourites",
      response: {
        200: favouritesListResponseSchema,
        401: errorEnvelopeSchema,
        429: errorEnvelopeSchema,
        500: errorEnvelopeSchema,
      },
      responseDocs: {
        200: { description: "The user's favourite locations, ordered as described above." },
        401: { description: "No valid session (`UNAUTHENTICATED`)." },
        429: { description: "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`." },
        500: { description: "Unexpected server error (`INTERNAL_ERROR`)." },
      },
    },
    async handler(request) {
      return favouritesService.listForUser(sessionUserId(request.userId));
    },
  });

  fastify.withTypeProvider<ZodTypeProvider>().route({
    method: "POST",
    url: "/favourites",
    schema: {
      tags: ["Favourites"],
      summary: "Save a favourite location",
      description:
        "Saves a resolved location (the `location` object from GET /api/v1/weather) as a " +
        "favourite, placed at the top of the list. Coordinates are the identity: saving " +
        "the same `lat`/`lon` twice responds 409, and each user can hold at most 20 " +
        "favourites. Requires a Better-Auth session cookie — see the Authentication page.",
      operationId: "addFavourite",
      body: favouriteCreateSchema,
      response: {
        201: favouriteItemSchema,
        400: errorEnvelopeSchema,
        401: errorEnvelopeSchema,
        409: errorEnvelopeSchema,
        429: errorEnvelopeSchema,
        500: errorEnvelopeSchema,
      },
      responseDocs: {
        201: { description: "The favourite was saved; the created row is returned." },
        400: {
          description:
            "Invalid body (`VALIDATION_ERROR`) or the 20-favourite cap was hit " +
            "(`FAVOURITES_LIMIT_REACHED`).",
        },
        401: { description: "No valid session (`UNAUTHENTICATED`)." },
        409: { description: "This location is already a favourite (`ALREADY_FAVOURITE`)." },
        429: { description: "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`." },
        500: { description: "Unexpected server error (`INTERNAL_ERROR`)." },
      },
    },
    async handler(request, reply) {
      const favourite = await favouritesService.add(sessionUserId(request.userId), request.body);
      reply.status(201);
      return favourite;
    },
  });

  fastify.withTypeProvider<ZodTypeProvider>().route({
    method: "PUT",
    url: "/favourites/order",
    schema: {
      tags: ["Favourites"],
      summary: "Reorder favourite locations",
      description:
        "Persists a manual display order. The body must contain every one of the " +
        "signed-in user's favourite ids exactly once, top first; each favourite's " +
        "`sortOrder` becomes its position in the list. If the id set does not match the " +
        "user's current favourites (for example the list changed in another tab) the " +
        "request is rejected with 409 `FAVOURITES_OUT_OF_SYNC` — re-fetch and retry. " +
        "Responds with the freshly ordered list. Requires a Better-Auth session cookie — " +
        "see the Authentication page.",
      operationId: "reorderFavourites",
      body: favouritesReorderSchema,
      response: {
        200: favouritesListResponseSchema,
        400: errorEnvelopeSchema,
        401: errorEnvelopeSchema,
        409: errorEnvelopeSchema,
        429: errorEnvelopeSchema,
        500: errorEnvelopeSchema,
      },
      responseDocs: {
        200: { description: "The order was saved; the full reordered list is returned." },
        400: {
          description:
            "Invalid body (`VALIDATION_ERROR`): empty, more than 20 entries, or duplicate ids.",
        },
        401: { description: "No valid session (`UNAUTHENTICATED`)." },
        409: {
          description:
            "The id set does not match the user's current favourites " +
            "(`FAVOURITES_OUT_OF_SYNC`) — re-fetch the list and retry.",
        },
        429: { description: "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`." },
        500: { description: "Unexpected server error (`INTERNAL_ERROR`)." },
      },
    },
    async handler(request) {
      return favouritesService.reorder(sessionUserId(request.userId), request.body.ids);
    },
  });

  fastify.withTypeProvider<ZodTypeProvider>().route({
    method: "DELETE",
    url: "/favourites/:id",
    schema: {
      tags: ["Favourites"],
      summary: "Remove a favourite location",
      description:
        "Removes one of the signed-in user's favourites. Favourites that do not exist — " +
        "or belong to another user — respond 404 (ownership is never revealed). " +
        "Requires a Better-Auth session cookie — see the Authentication page.",
      operationId: "deleteFavourite",
      params: favouriteDeleteParamsSchema,
      response: {
        // 204: fastify sends no body for 204 responses.
        204: z.null(),
        400: errorEnvelopeSchema,
        401: errorEnvelopeSchema,
        404: errorEnvelopeSchema,
        429: errorEnvelopeSchema,
        500: errorEnvelopeSchema,
      },
      responseDocs: {
        204: { description: "The favourite was removed. No body." },
        400: { description: "Invalid request (`VALIDATION_ERROR`): bad `id` value." },
        401: { description: "No valid session (`UNAUTHENTICATED`)." },
        404: { description: "No such favourite for this user (`NOT_FOUND`)." },
        429: { description: "Rate limit exceeded (`RATE_LIMITED`) — see `retry-after`." },
        500: { description: "Unexpected server error (`INTERNAL_ERROR`)." },
      },
    },
    async handler(request, reply) {
      const deleted = await favouritesService.deleteOwned(
        sessionUserId(request.userId),
        request.params.id,
      );
      if (!deleted) {
        // Missing and not-owned are indistinguishable on purpose: never
        // reveal whether an id exists for another user.
        throw new AppError(404, ErrorCodes.NOT_FOUND, "Favourite not found.");
      }
      reply.status(204).send(null);
    },
  });
}